Add support for additional ENV variables injection. Overall tasks clean-up.
