+++ .travis.yml
@@ -15,6 +15,7 @@ env:
     - ANSIBLE=2.3
     - ANSIBLE=2.4
     - ANSIBLE=2.5
+    - ANSIBLE=2.6
 
 # Install tox
 install:
