+++ defaults/main.yml
@@ -4,30 +4,34 @@
 minio_server_bin: /usr/local/bin/minio
 minio_client_bin: /usr/local/bin/mc
 
-# Runtime user and group for the minio server service
+# Runtime user and group for the Minio server service
 minio_user: minio
 minio_group: minio
 
-# Path to the file containing the ENV variables for the minio server
+# Path to the file containing the ENV variables for the Minio server
 minio_server_envfile: /etc/default/minio
 
 # Minio server listen address
 minio_server_addr: ":9091"
 
 # Minio server data directories
-minio_server_datadirs: [ ]
+minio_server_datadirs:
+  - /var/lib/minio
 minio_server_make_datadirs: true
 
 # Minio server cluster node list.
 minio_server_cluster_nodes: [ ]
 
-# Additional minio server CLI options
+# Additional environment variables to be set in Minio server environment
+minio_server_env_extra: ""
+
+# Additional Minio server CLI options
 minio_server_opts: ""
 
 # Minio access and secret keys
 minio_access_key: ""
 minio_secret_key: ""
 
-# Switches to enable/disable the minio server and/or minio client installation.
+# Switches to enable/disable the Minio server and/or Minio client installation.
 minio_install_server: true
 minio_install_client: true
