--- molecule/alternative/molecule.yml
@@ -1,50 +1,0 @@
----
-driver:
-  name: docker
-lint:
-  name: yamllint
-
-platforms:
-  - name: minio-centos-7
-    image: paulfantom/centos-molecule:7
-    privileged: true
-    volumes:
-      - /sys/fs/cgroup:/sys/fs/cgroup:ro
-  - name: minio-fedora-27
-    image: paulfantom/fedora-molecule:27
-    privileged: true
-    volumes:
-      - /sys/fs/cgroup:/sys/fs/cgroup:ro
-  - name: minio-debian-9
-    image: paulfantom/debian-molecule:9
-    privileged: true
-    volumes:
-      - /sys/fs/cgroup:/sys/fs/cgroup:ro
-  - name: minio-ubuntu-18.04
-    image: paulfantom/ubuntu-molecule:18.04
-    privileged: true
-    volumes:
-      - /sys/fs/cgroup:/sys/fs/cgroup:ro
-  - name: minio-ubuntu-16.04
-    image: paulfantom/ubuntu-molecule:16.04
-    privileged: true
-    volumes:
-      - /sys/fs/cgroup:/sys/fs/cgroup:ro
-
-provisioner:
-  name: ansible
-  lint:
-    name: ansible-lint
-  playbooks:
-    create: ../default/create.yml
-    prepare: ../default/prepare.yml
-    converge: playbook.yml
-    destroy: ../default/destroy.yml
-
-scenario:
-  name: alternative
-verifier:
-  name: testinfra
-  lint:
-    name: flake8
-    enabled: true
