--- molecule/default/prepare.yml
@@ -1,5 +1,0 @@
----
-- name: Prepare
-  hosts: all
-  gather_facts: false
-  tasks: []
