+++ molecule/resources/destroy.yml
