+++ molecule/resources/tests/test_minio_common.py
@@ -1,0 +1,13 @@
+import os
+import testinfra.utils.ansible_runner
+
+
+testinfra_hosts = testinfra.utils.ansible_runner.AnsibleRunner(
+    os.environ['MOLECULE_INVENTORY_FILE']).get_hosts('all')
+
+
+def test_minio_service(host):
+
+  s = host.service('minio')
+  assert s.is_running
+  assert s.is_enabled
