--- tasks/client.yml
@@ -1,13 +1,0 @@
----
-- name: download minio client
-  get_url:
-    url: "{{ minio_client_download_url }}"
-    dest: "{{ minio_client_bin }}"
-    owner: "root"
-    group: "root"
-    mode: 0755
-    checksum: "sha256:{{ minio_client_checksum }}"
-  register: _download_client
-  until: _download_client is succeeded
-  retries: 5
-  delay: 2
