--- tasks/preflight.yml
@@ -1,22 +1,0 @@
----
-- block:
-    - name: Compose Minio download url
-      set_fact:
-        minio_server_download_url: "https://dl.minio.io/server/minio/release/linux-{{ go_arch_map[ansible_architecture] | default(ansible_architecture) }}/minio"
-
-    - name: "Minio server checksum for {{ go_arch_map[ansible_architecture] | default(ansible_architecture) }} architecture"
-      set_fact:
-        minio_server_checksum: "{{ lookup('url', minio_server_download_url + '.sha256sum').split(' ')[0] }}"
-  when:
-    - minio_install_server
-
-- block:
-    - name: Compose MC download url
-      set_fact:
-        minio_client_download_url: "https://dl.minio.io/client/mc/release/linux-{{ go_arch_map[ansible_architecture] | default(ansible_architecture) }}/mc"
-
-    - name: "Minio client checksum for {{ go_arch_map[ansible_architecture] | default(ansible_architecture) }} architecture"
-      set_fact:
-        minio_client_checksum: "{{ lookup('url', minio_client_download_url + '.sha256sum').split(' ')[0] }}"
-  when:
-    - minio_install_client
