+++ tasks/python_sni.yml
@@ -1,4 +1,5 @@
 ---
+
 - name: install python-pip and SNI support packages
   package:
     name: "{{ item }}"
