+++ templates/minio.env.j2
@@ -17,3 +17,5 @@ MINIO_ACCESS_KEY="{{ minio_access_key }}"
 # Secret key of the server.
 MINIO_SECRET_KEY="{{ minio_secret_key }}"
 {% endif %}
+
+{{ minio_server_env_extra }}
