+++ tox.ini
@@ -1,6 +1,6 @@
 [tox]
 minversion = 1.8
-envlist = py{27}-ansible{23,24,25}
+envlist = py{27}-ansible{23,24,25,26}
 skipsdist = true
 
 [travis:env]
@@ -8,6 +8,7 @@ ANSIBLE=
   2.3: ansible23
   2.4: ansible24
   2.5: ansible25
+  2.6: ansible26
 
 [testenv]
 passenv = *
@@ -16,5 +17,6 @@ deps =
     ansible23: ansible<2.4
     ansible24: ansible<2.5
     ansible25: ansible<2.6
+    ansible26: ansible<2.7
 commands =
     {posargs:molecule test --all --destroy always}
